Update README.md for local setup
Just tried to setup the npm stuff after freshly cloning the repository and it didn't work. The old instructions had you do everything in the /app directory, but now package.json is no longer found there so I removed the instruction that told you to be in /app. Let me know if this is incorrect, it is just what worked for me, and it should be figured out for if the graders want to run the code locally